refactor: move map tests
